pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "licenseguard"
# NOTE: Keep in sync with licenseguard.__init__.__version__
version = "0.3.0"
description = "Scan installed Python packages for license compliance against a requirements file"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "LicenseGuard contributors" }]
keywords = ["license", "compliance", "dependencies", "requirements"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "packaging>=23.0",
    "PyYAML>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "httpx>=0.24.0,<0.28.0"]

[project.scripts]
licenseguard = "licenseguard.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
